move proxies out of MTO2 so i can use them in auto
